# Conflicts:
#	codes/structure_model/gstructure
#	codes/structure_model/run_galaxy.c
#	codes/structure_model/sample_gstructure
#	codes/structure_model/test_mcmc
#	codes/structure_model/variables.h
#	run/PARAMETER

+++ codes/structure_model/variables.h
@@ -21,7 +21,7 @@ extern int Mah_simu;
 extern int Do_preheating;
 extern int Do_reinfall;
 extern int Star_formation_model;
-extern float Mass_bin;
+extern double Mass_bin;
 extern double Redshift_end;
 
 extern int Resize_radius_bins;
